fix(airfield): fix rounding errors on airfield coordinates
Using doubles was making them go skew-whiff and breaking the odd test

+++ app/Models/Airfield/Airfield.php
@@ -45,8 +45,8 @@ class Airfield extends Model implements MinStackDataProviderInterface
 
     protected $casts = [
         'groundspeed' => 'integer',
-        'latitude' => 'float',
-        'longitude' => 'float',
+        'latitude' => 'decimal',
+        'longitude' => 'decimal',
     ];
 
     /**
